Moved introductory offer utility functions to single file
